Refactor states
